Generalize migration runner
